refactor: order package by part
