+++ _1iterator/Book.kt
@@ -1,3 +1,3 @@
-package iterator
+package _1iterator
 
 class Book(val name: String)
