+++ _1iterator/Iterable.kt
@@ -1,4 +1,4 @@
-package iterator
+package _1iterator
 
 interface Iterable<E> {
     fun iterator(): Iterator<E>
