+++ _2adapter/adapter_inheritance/Print.kt
@@ -1,4 +1,4 @@
-package adapter.adapter_inheritance
+package _2adapter.adapter_inheritance
 
 interface Print {
     fun printWeak()
