+++ _2adapter/adapter_inheritance/PrintBanner.kt
@@ -1,4 +1,4 @@
-package adapter.adapter_inheritance
+package _2adapter.adapter_inheritance
 
 class PrintBanner(string: String) : Banner(string), Print {
     override fun printWeak() {
