+++ _2adapter/adpater_delegation/Print.kt
@@ -1,4 +1,4 @@
-package adapter.adpater_delegation
+package _2adapter.adpater_delegation
 
 abstract class Print {
     abstract fun printWeak()
