+++ _2adapter/adpater_delegation/main.kt
@@ -1,4 +1,4 @@
-package adapter.adpater_delegation
+package _2adapter.adpater_delegation
 
 fun main() {
     val printBanner = PrintBanner(Banner("Hello"))
